Move all classes that describe display settings to the new directory
